Этап 2. Бэкенд - Чеклист. Исправлены контроллеры для getUser и getArticles, в отдельные файлы вынесены обработчик ошибок и rate Limiter
