+++ app.js
@@ -1,21 +1,16 @@
 require('dotenv').config();
 const express = require('express');
 const helmet = require('helmet');
-const rateLimit = require('express-rate-limit');
 const mongoose = require('mongoose');
 const bodyParser = require('body-parser');
 const cookieParser = require('cookie-parser');
-const { isCelebrate } = require('celebrate');
-const { requestLogger, errorLogger } = require('./middlewares/logger');
 
-const limiter = rateLimit({
-  windowMs: 15 * 60 * 1000, // за 15 минут
-  max: 100, // можно совершить максимум 100 запросов с одного IP
-});
+const { requestLogger, errorLogger } = require('./middlewares/logger');
+const limiter = require('./rateLimiter');
 const {
   PORT, DATABASE_URL,
 } = require('./config');
-
+const errorsHandler = require('./errorsHandler');
 const routes = require('./routes');
 
 mongoose.set('useUnifiedTopology', true);
@@ -38,10 +33,6 @@ app.use(requestLogger);
 app.use(routes);
 app.use(errorLogger);
 
-app.use((err, req, res, next) => {
-  const statusCode = isCelebrate(err) ? 400 : err.statusCode;
-  res.status(statusCode || 500).send({ message: statusCode === 500 ? 'На сервере произошла ошибка' : err.message });
-  return next();
-});
+app.use(errorsHandler);
 
 app.listen(PORT);
